Switch to pinia store (#5)
