+++ src/main.ts
@@ -1,5 +1,6 @@
 import { createApp } from 'vue'
-import { store, key } from './store'
+import { createPinia } from 'pinia'
+import piniaPluginPersistedState from 'pinia-plugin-persistedstate'
 import router from './router'
 import i18n from './i18n'
 import App from './App.vue'
@@ -7,8 +8,11 @@ import App from './App.vue'
 import 'bootstrap/dist/css/bootstrap.min.css'
 import 'bootstrap'
 
+const pinia = createPinia()
+  .use(piniaPluginPersistedState)
+
 createApp(App)
-  .use(store, key)
+  .use(pinia)
   .use(router)
   .use(i18n)
-  .mount('#app');
+  .mount('#app')
