+++ src/util/getPlayerColorCode.ts
@@ -14,7 +14,7 @@ export default function(playerColor: PlayerColor) : string {
     case PlayerColor.ORANGE:
       return "#f08418"
     case PlayerColor.RED:
-      return "#b41917";
+      return "#b41917"
     default:
       throw new Error(`Invalid player color: ${playerColor}.`)
   }
