+++ tests/unit/helper/mockLunaState.ts
@@ -1,6 +1,6 @@
-import DifficultyLevel from '@/services/enum/DifficultyLevel';
-import LunaState from '@/services/LunaState';
-import mockCardDeck, { MockCardDeckParams } from './mockCardDeck';
+import DifficultyLevel from '@/services/enum/DifficultyLevel'
+import LunaState from '@/services/LunaState'
+import mockCardDeck, { MockCardDeckParams } from './mockCardDeck'
 
 export default function (params?: MockLunaStateParams) : LunaState {  
   let cardDeckParams = params?.cardDeck
